updating to detect need to clean cover data more
note that cleaning needs to be done more before interpreting final results.
For KBS: There are several singletons (species that are only observed once or twice), some that may be mis-IDed, and some that are only first noticed in July or August, which is hard to believe. For now they should be removed but the cleaning on this should take place in plant_comp_clean_L0.R.
Need to make some decisions about which of these species to include or re-assign to a different species. All of that needs to be done in plant_comp_clean_L0.R

UMBS still needs to be checked like this also.